+++ src/main/java/io/supertokens/webserver/api/webauthn/CredentialsRegisterAPI.java
@@ -22,6 +22,8 @@ import io.supertokens.pluginInterface.Storage;
 import io.supertokens.pluginInterface.exceptions.StorageQueryException;
 import io.supertokens.pluginInterface.multitenancy.TenantIdentifier;
 import io.supertokens.pluginInterface.multitenancy.exceptions.TenantOrAppNotFoundException;
+import io.supertokens.pluginInterface.webauthn.exceptions.DuplicateCredentialException;
+import io.supertokens.pluginInterface.webauthn.exceptions.UserIdNotFoundException;
 import io.supertokens.pluginInterface.webauthn.exceptions.WebauthNOptionsNotExistsException;
 import io.supertokens.webauthn.WebAuthN;
 import io.supertokens.webauthn.data.WebauthNCredentialResponse;
@@ -58,6 +60,7 @@ public class CredentialsRegisterAPI extends WebserverAPI {
             String webauthnGeneratedOptionsId = InputParser.parseStringOrThrowError(input, "webauthnGeneratedOptionsId", false);
             JsonObject credentialsData = InputParser.parseJsonObjectOrThrowError(input, "credential", false);
 
+            // useridmapping is not required to be called separately here because it is handled in the underlying parts
             WebauthNCredentialResponse savedCredential = WebAuthN
                     .registerCredentials(storage, tenantIdentifier, recipeUserId,
                             webauthnGeneratedOptionsId, credentialsData);
@@ -92,6 +95,14 @@ public class CredentialsRegisterAPI extends WebserverAPI {
         } catch (WebauthNOptionsNotExistsException e) {
             JsonObject result = new JsonObject();
             result.addProperty("status", "OPTIONS_NOT_FOUND_ERROR");
+            sendJsonResponse(200, result, resp);
+        } catch (DuplicateCredentialException e) {
+            JsonObject result = new JsonObject();
+            result.addProperty("status", "CREDENTIAL_ALREADY_EXISTS_ERROR");
+            sendJsonResponse(200, result, resp);
+        } catch (UserIdNotFoundException e) {
+            JsonObject result = new JsonObject();
+            result.addProperty("status", "UNKNOWN_USER_ID_ERROR");
             sendJsonResponse(200, result, resp);
         }
     }