+++ src/test/java/io/supertokens/test/emailpassword/UpdateUsersEmailAndPasswordTest.java
@@ -18,6 +18,7 @@ package io.supertokens.test.emailpassword;
 
 import io.supertokens.Main;
 import io.supertokens.emailpassword.EmailPassword;
+import io.supertokens.emailverification.EmailVerification;
 import io.supertokens.pluginInterface.STORAGE_TYPE;
 import io.supertokens.pluginInterface.authRecipe.AuthRecipeUserInfo;
 import io.supertokens.pluginInterface.emailpassword.exceptions.DuplicateEmailException;
@@ -153,6 +154,39 @@ public class UpdateUsersEmailAndPasswordTest {
 
             Assert.assertEquals(userInfo.getSupertokensUserId(), changedCredentialsUserInfo.getSupertokensUserId());
             Assert.assertEquals("dave.doe@example.com", changedCredentialsUserInfo.loginMethods[0].email);
+            Assert.assertFalse(changedCredentialsUserInfo.loginMethods[0].verified);
+        });
+    }
+
+    @Test
+    public void testUpdateEmailAndPasswordFromVerifiedToNotVerified() throws Exception {
+        TestingProcessManager.withProcess(process -> {
+            Main main = process.getProcess();
+
+            if (StorageLayer.getStorage(main).getType() != STORAGE_TYPE.SQL) {
+                return;
+            }
+
+            // given
+            AuthRecipeUserInfo userInfo = EmailPassword.signUp(main, "john.doe@example.com", "password");
+
+            String emailVerificationToken = EmailVerification.generateEmailVerificationToken(main, userInfo.getSupertokensUserId(),
+                    userInfo.loginMethods[0].email);
+            EmailVerification.verifyEmail(main, emailVerificationToken);
+            AuthRecipeUserInfo originalUser = EmailPassword.signIn(main, "john.doe@example.com", "password");
+            Assert.assertTrue(originalUser.loginMethods[0].verified);
+
+            // when
+            EmailPassword.updateUsersEmailOrPassword(main, userInfo.getSupertokensUserId(), "dave.doe@example.com",
+                    "newPassword");
+
+            // then
+            AuthRecipeUserInfo changedCredentialsUserInfo = EmailPassword.signIn(main, "dave.doe@example.com",
+                    "newPassword");
+
+            Assert.assertEquals(userInfo.getSupertokensUserId(), changedCredentialsUserInfo.getSupertokensUserId());
+            Assert.assertEquals("dave.doe@example.com", changedCredentialsUserInfo.loginMethods[0].email);
+            Assert.assertFalse(changedCredentialsUserInfo.loginMethods[0].verified);
         });
     }
 }