+++ src/test/java/io/supertokens/test/webauthn/WebAuthNFlowTest.java
@@ -106,5 +106,4 @@ public class WebAuthNFlowTest {
         assertEquals(signInResponse, ePSignInResponse);
     }
 
-
 }