+++ CHANGELOG.md
@@ -11,6 +11,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Fixes `StorageTransactionLogicException` in bulk import when not using userRoles and totpDevices in import json.
 - MFA only required in Bulk Import if it's used in input data
+- Fixes issue with reloading all resources when exception occurs while loading a resource, other valid resources were offloaded from the memory. Now we log the exception and continue loading other resources.
+- Adds `USE_STRUCTURED_LOGGING` environment variable to control the logging format.
 
 ## [10.0.2]