+++ CHANGELOG.md
@@ -25,6 +25,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   - PUT `/recipe/webauthn/user/email`
   - DELETE `/recipe/webauthn/user/credential/remove`
   - DELETE `/recipe/webauthn/options/remove`
+- Adds additional indexing for `emailverification_verified_emails`
+- 
 
 ### Migration