+++ CHANGELOG.md
@@ -192,6 +192,13 @@ CREATE INDEX webauthn_account_recovery_token_expires_at_index ON webauthn_accoun
 CREATE INDEX webauthn_account_recovery_token_email_index ON webauthn_account_recovery_tokens (app_id, tenant_id, email);
 ```
 
+## [10.0.3]
+
+- Fixes `StorageTransactionLogicException` in bulk import when not using userRoles and totpDevices in import json.
+- MFA only required in Bulk Import if it's used in input data
+- Fixes issue with reloading all resources when exception occurs while loading a resource, other valid resources were offloaded from the memory. Now we log the exception and continue loading other resources.
+- Adds `USE_STRUCTURED_LOGGING` environment variable to control the logging format.
+
 ## [10.0.2]
 
 - Fixes `NullPointerException` in user search API.